Implementing product list and detail workflow with cart events
Add to cart button is wired up to an event. This happens now
via a console log statement. In addition, the product detail
has been expanded to include an image. The main app controller
has the add to cart method, but done not activate anything.
In addition, the controller now includes a variable for
cart data.